Move scm-1 rockspec to project root
There is no much profit in storing release rockspecs in the rockspecs
directory. It may be confusing to have rockspecs from previous versions
in the latest version of repo.

Rockspecs are stored in version control under the git tag anyway. Both
it is better to publish all new releases into tarantool/rocks repo