Push errors from rd_kafka_consumer_poll to error queue
Before this patch, if there was any error during the execution of rd_kafka_consumer_poll,
then the error was ignored and the function was run again.
As a result, the loop silently became infinite and used 100% CPU.

Now the errors received when calling rd_kafka_consumer_poll
are sent to the error queue and you can see what errors occur.