set a name to kafka threads
Recently we faced some strange issue in one project. There were a
lot of different records in metrics about threads stats. It was
impossible to understand where all this threads appeared.
This patch adds names to kafka threads. So it will be possible to
count amount of threads that kafka module creates.