Use ExternalProject to build librdkafka
We need to provide infrustructure for applying patches in
kafka library. The most reliable and convenient way to do it
is using ExternalProject cmake module. Moreover, it's used
in main tarantool repository.

So let's be consistent and apply patches with ExternalProject
module. This commit replaces old custom_command with it.