patch: fix deref after NULL in rdkafka_assignment
In `rd_kafka_assignment_handle_OffsetFetch` if `err` argument
was set to a non-zero value different from `RD_KAFKA_RESP_ERR__DESTROY`
and `RD_KAFKA_RESP_ERR__IN_PROGRESS`, and if allow_retry is true,
then offset won't be set by `rd_kafka_handle_OffsetFetch` and we'll get
NULL dereference in error throwing, applying and destroying `offsets`.

As it looks vital to throw an error anyway, let's just throw it
with incorrect number of partition (-1). As for applying and destroying,
let's just check if `offsets` is not NULL.

Closes tarantool/security#47